Allow scanning from any page using scanner device
